rename FastDetector->Detector
