+++ feat/tests/performance_testing.py
@@ -1,5 +1,5 @@
 # %%
-from feat.FastDetector import MPDetector
+from feat.MPDetector import MPDetector
 import os
 from feat.utils.io import get_test_data_path
 import cProfile
@@ -7,7 +7,6 @@ import pstats
 
 multi_face = os.path.join(get_test_data_path(), "multi_face.jpg")
 
-# detector = FastDetector()
 detector = MPDetector(device="mps", emotion_model="resmasknet", identity_model="facenet")
 
 # detector.detect(multi_face, data_type='image')
